+++ src/constants/variables.ts
@@ -1,0 +1,12 @@
+export const VARIABLES = {
+  githubToken: "github-token",
+  reportPath: "report-path",
+  issuePrefix: "issue-prefix",
+  issueLabels: "issue-labels",
+  issueAssignees: "issue-assignees",
+  issueFooter: "issue-footer",
+  addProjectLabel: "add-project-label",
+  addComment: "add-comment",
+  jobSummary: "job-summary",
+  quite: "quite",
+};