+++ src/models/index.ts
@@ -1,3 +1,4 @@
+export * from "./FailedTestInfo";
 export * from "./JSONReport";
 export * from "./TestError";
 export * from "./TestStatus";