+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.0]
+
+- Added the `issue-assignees` input to assign the issue to specific users.
+- Fix in summary title
+- Fix in summary links
+
 ## [1.0.0]
 
 - Initial release